refactor: color palette
